fix: move team prompt handling in createNewMeetingPhase
