+++ packages/server/database/types/TeamPromptResponsesPhase.ts
@@ -2,11 +2,16 @@ import GenericMeetingPhase from './GenericMeetingPhase'
 import TeamPromptResponseStage from './TeamPromptResponseStage'
 
 export default class TeamPromptResponsesPhase extends GenericMeetingPhase {
-  stages: TeamPromptResponseStage[]
+  stages: [TeamPromptResponseStage, ...TeamPromptResponseStage[]]
   phaseType!: 'RESPONSES'
 
   constructor(teamMemberIds: string[]) {
     super('RESPONSES')
-    this.stages = teamMemberIds.map((teamMemberId) => new TeamPromptResponseStage({teamMemberId}))
+    if (teamMemberIds.length < 1) {
+      throw new Error('TeamPromptResponsesPhase must have at least one team member')
+    }
+    this.stages = teamMemberIds.map(
+      (teamMemberId) => new TeamPromptResponseStage({teamMemberId})
+    ) as [TeamPromptResponseStage, ...TeamPromptResponseStage[]]
   }
 }
